Add image to Betscreen
  * Image was added to bet page, which is fetched from the respective
bet data file.
  * SVG is now handled using react-native-svg. A wrapper component
creates SVG React components that can be manipulated to our needs. An
additional wrapper contains all the possible Icons in our object, and
provides these as SVGs wrapped from the former component.
  * The local JSON data was reorganized. Now there is a general file for
the homescreen, and specific files for each individual betscreen, as
would be in a remote DB.
  * The data is fetched locally, via a wrapper staticCache. This is due
to inability to have dynamic requires. Later when the remote is added,
this can be removed.
  * Made StatusBar permanent